+++ src/Engine/GameManager.h
@@ -24,10 +24,10 @@ public:
 	}
 
 	float64 fixedUpdateDT = 1.f / FixedUpdateExecution;
-	float32 accDt = 0.f;
+	float64 accDt = 0.f;
 
-	int16 fpsCount = 0.f;
-	float32 fpsTimer = 0.f;
+	int16 fpsCount = 0;
+	float64 fpsTimer = 0.f;
 
 	GameManager() = default;
